gestionCommande

+++ resources/views/users/index.blade.php
@@ -7,34 +7,30 @@
     <div class="content-wrapper">
         <!-- Content Header (Page header) -->
         <div class="content-header">
-            <div class="container-fluid">
-                <div class="row mb-2">
+                <div class="container-fluid">
+                    <div class="row mb-2">
                     <div class="col-sm-6">
-                        <h1 class="m-0 text-dark">ACCUEIL</h1>
+                        <h1 class="m-0 text-info">GESTION DES UTILISATEURS</h1>
                     </div><!-- /.col -->
                     <div class="col-sm-6">
                         <ol class="breadcrumb float-sm-right">
-                            <li class="breadcrumb-item"><a href="#">Accueil</a></li>
-                            <li class="breadcrumb-item active">BIENVENU</li>
+                        <li class="breadcrumb-item"><a href="{{ route('users.create') }}" role="button" class="btn btn-primary">AJOUTER UN UTILISATEUR</a></li>
+                        <li class="breadcrumb-item"><a href="{{ route('roles.index') }}" role="button" class="btn btn-primary">RÔLES</a></li>
+                        <li class="breadcrumb-item active"><a href="{{ route('permissions.index') }}" role="button" class="btn btn-primary">PERMISSIONS</a></li>
                         </ol>
                     </div><!-- /.col -->
-                </div><!-- /.row -->
-            </div><!-- /.container-fluid -->
-        </div>
+                    </div><!-- /.row -->
+                </div><!-- /.container-fluid -->
+    </div>
         <!-- /.content-header -->
 
         <!-- Main content -->
-        <section class="content">
-            <div class="container-fluid">
-                <div class="row">
-                    <div class="col-lg-10 col-lg-offset-1">
-                        <h1><i class="fa fa-users"></i> User Administration <a href="{{ route('roles.index') }}" class="btn btn-default pull-right">Roles</a>
-                            <a href="{{ route('permissions.index') }}" class="btn btn-default pull-right">Permissions</a></h1>
-                        <hr>
-                        <div class="table-responsive">
-                            <table class="table table-bordered table-striped">
-
-                                <thead>
+        <div class="col-12">
+            <div class="card border-danger border-0">
+                <div class="card-header bg-info text-center">LISTE D'ENREGISTREMENT DES UTILISATEURS</div>
+                    <div class="card-body">
+                        <table id="example1" class="table table-bordered table-responsive-md table-striped text-center">
+                            <thead>
                                 <tr>
                                     <th>Name</th>
                                     <th>Email</th>
@@ -42,7 +38,7 @@
                                     <th>User Roles</th>
                                     <th>Operations</th>
                                 </tr>
-                                </thead>
+                            </thead>
 
                                 <tbody>
                                 @foreach ($users as $user)
@@ -53,10 +49,11 @@
                                         <td>{{ $user->created_at->format('F d, Y h:ia') }}</td>
                                         <td>{{  $user->roles()->pluck('name')->implode(' ') }}</td>{{-- Retrieve array of roles associated to a user and convert to string --}}
                                         <td>
-                                            <a href="{{ route('users.edit', $user->id) }}" class="btn btn-info pull-left" style="margin-right: 3px;">Edit</a>
 
-                                            {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id] ]) !!}
-                                            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+
+                                            <a href="{{ route('users.edit', $user->id) }}" role="button" class="btn btn-primary"><i class="fas fa-edit"></i></a>
+                                            {!! Form::open(['method' => 'DELETE', 'route'=>['users.destroy', $user->id], 'style'=> 'display:inline', 'onclick'=>"if(!confirm('Êtes-vous sûr de vouloir supprimer cet enregistrement ?')) { return false; }"]) !!}
+                                            <button class="btn btn-danger"><i class="far fa-trash-alt"></i></button>
                                             {!! Form::close() !!}
 
                                         </td>
@@ -67,7 +64,6 @@
                             </table>
                         </div>
 
-                        <a href="{{ route('users.create') }}" class="btn btn-success">Add User</a>
 
                     </div>
                 </div>
